[android] rename HelloForms.Android -> HelloForms.Droid
The Xamarin.Forms templates create `Foo.Droid.csproj` for Android head
projects instead of `Foo.Android.csproj`.

The reason we do this is because a top level namespace like:

    namespace HelloForms.Android
    {
        public class MainActivity : Xamarin.Forms.Platform.Android.FormsAppCompatActivity
        {
            protected override void OnCreate(Bundle savedInstanceState)
            {
                // ...
            }
        }
    }

This conflicts with `Android` namespaces, that are normally listed via
`using` statements at the top, such as:

    using Android.App;
    using Android.Content.PM;
    using Android.Runtime;
    using Android.Views;
    using Android.Widget;
    using Android.OS;

If you tried to use `Android.Views.View` in the above `OnCreate`
method, you get an error such as:

    error CS0234: The type or namespace name 'Views' does not exist in the namespace 'HelloForms.Android'

To avoid this problem, I think we should go ahead and name the project
`HelloForms.Droid.csproj`.