Bump to Microsoft.NET.Workload.Android 11.0.100.209
Context: https://github.com/xamarin/xamarin-android/pull/5236

Fixes problem with Fast Deployment on Android caused by incorrect
paths in `@(ResolvedFilesToPublish)` from the dotnet/sdk.

Hoping it might help the random iOS failure, too? We saw things in the
build log such as:

    asset directory did not exist: /Users/runner/work/1/s/HelloAndroid/obj/Debug/net5.0-android/assets/ (TaskId:253)
    Executing link --manifest /Users/runner/work/1/s/HelloAndroid/obj/Debug/net5.0-android/android/manifest/AndroidManifest.xml [...]
    ILLINK : error MT2301: The linker step 'Setup' failed during processing. [/Users/runner/work/1/s/HelloiOS/HelloiOS.csproj]

`net6-samples.sln` will build projects in parallel, so we're not sure
if the Android log message is affecting the iOS build or not.