Update iOS sample to be visible in dark mode (#69)
When I ran the iOS sample initially, I thought it didn't work because the entire screen was white. Because the device was already in dark mode, the status bar and the label were rendered white. This might be confusing to developers, since this is a preview, and things not working is expected.

I changed the sample to use the system background color if that API is available. To keep changes minimal I just set the label to blue, which isn't ideal but at least is visible.

I tested with the simulator for iOS 14 and iOS 11.